Fearure/product

+++ ShopOnline.Application/Command/Products/ManagementProductService.cs
@@ -7,6 +7,7 @@ using ShopOnline.Data.Entities;
 using ShopOnline.Utill;
 using ShopOnlineViewModel.Catalog.Product;
 using ShopOnlineViewModel.Catalog.Product.Manage;
+using ShopOnlineViewModel.Catalog.ProductImage;
 using ShopOnlineViewModel.Common;
 using System;
 using System.Collections.Generic;
@@ -71,7 +72,7 @@ namespace ShopOnline.Application.Command.Products
                         DateCreate = DateTime.Now,
                         FileSize = request.ThumbnailImange.Length,
                         ImagePath= await this.SaveFile(request.ThumbnailImange),
-                        IsDefaul=true,
+                        IsDefault=true,
                         SortOder=1,
                     }
                 };
@@ -99,7 +100,7 @@ namespace ShopOnline.Application.Command.Products
 
             if (request.ThumbnailImange != null)
             {
-                var ThumbnailImange = await _context.ProductImages.FirstOrDefaultAsync(i => i.IsDefaul == true && i.ProdutId == request.Id);
+                var ThumbnailImange = await _context.ProductImages.FirstOrDefaultAsync(i => i.IsDefault == true && i.ProductId == request.Id);
                
                 if (ThumbnailImange != null)
                 {
@@ -108,7 +109,7 @@ namespace ShopOnline.Application.Command.Products
                     ThumbnailImange.DateCreate = DateTime.Now;
                     ThumbnailImange.FileSize = request.ThumbnailImange.Length;
                     ThumbnailImange.ImagePath = await this.SaveFile(request.ThumbnailImange);
-                    ThumbnailImange.IsDefaul = true;
+                    ThumbnailImange.IsDefault = true;
                     ThumbnailImange.SortOder = 1;
                     _context.ProductImages.Update(ThumbnailImange);
                 }
@@ -119,8 +120,12 @@ namespace ShopOnline.Application.Command.Products
         public async Task<int> Delete(int productId)
         {
             var product = await _context.Products.FindAsync(productId);
+            if (product == null)
+            {
+                throw new ShopOnlineException($"can not product {productId}");
+            }
 
-            var images = await _context.ProductImages.FirstOrDefaultAsync(i=> i.ProdutId == product.Id);
+            var images =  _context.ProductImages.Where(i=> i.ProductId == productId);
            foreach( var image in images)
             {
                 _storageService.DeleteAsync(image.ImagePath);
@@ -216,6 +221,49 @@ namespace ShopOnline.Application.Command.Products
             return await _context.SaveChangesAsync() > 0;
         }
 
-      
+        public async Task<int> AddImange(int productId, ProductImageCreateRequest request)
+        {
+            var productImage = new ProductImage()
+            {
+                Caption = request.Caption,
+                DateCreate = DateTime.Now,
+                IsDefault = request.IsDefault,
+                ProductId = productId,
+                SortOder = request.SortOrder
+            };
+            if (request.ImageFile != null)
+            {
+                productImage.ImagePath = await this.SaveFile(request.ImageFile);
+                productImage.FileSize = request.ImageFile.Length;
+            }
+            _context.ProductImages.Add(productImage);
+            await _context.SaveChangesAsync();
+            return productImage.Id;
+        }
+
+        public Task<int> UpdateImage(int ImageId, string caption, bool isDefault)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task<int> RemoveImage(int imageId)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task<List<ProductImageViewModel>> GetListImage(int ProductIsd)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task<int> AddImange(int productId, List<IFormFile> files)
+        {
+            throw new NotImplementedException();
+        }
+
+        Task<List<ProductImageViewModel>> IProductManagementService.GetListImage(int ProductIsd)
+        {
+            throw new NotImplementedException();
+        }
     }
 }
